util: Move packet trace file read to protolib
This patch moves the code for opening an input protobuf packet trace into
a function defined in the protobuf library. This is because the code is
commonly used in decode scripts and is independent of the src protobuf
message.
(transplanted from 82773ace39fa84456493d3af1d0fa6c445e0f3e0)